requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
xero-python>=9.0.0
sqlalchemy>=2.0.36
psycopg2-binary>=2.9.9
python-dotenv==1.0.0
cryptography==41.0.7
requests==2.31.0
gunicorn==21.2.0
anthropic>=0.40.0
pyyaml>=6.0
cachetools>=5.3.0
pandas>=2.0.0
openpyxl>=3.1.0
redis>=5.0.0
